feat: port verify-tombstone command
